kpkg: removeInternal: add group support

+++ kpkg/modules/removeInternal.nim
@@ -10,6 +10,7 @@ proc removeInternal*(package: string, root = "",
         ignoreReplaces = false) =
 
     let init = getInit(root)
+    let pkg = parse_runfile(installedDir&"/"&actualPackage)
 
     if dirExists(installedDir&"/"&package&"-"&init):
         removeInternal(package&"-"&init, root, installedDir, ignoreReplaces)
@@ -21,18 +22,19 @@ proc removeInternal*(package: string, root = "",
     else:
         actualPackage = package
 
-    if not fileExists(installedDir&"/"&actualPackage&"/list_files"):
-        err("package "&package&" is not installed", false)
 
-    for line in lines installedDir&"/"&actualPackage&"/list_files":
-        discard tryRemoveFile(root&"/"&line)
+    if not pkg.isGroup:
+        if not fileExists(installedDir&"/"&actualPackage&"/list_files"):
+            err("package "&package&" is not installed", false)
 
-        if isEmptyOrWhitespace(toSeq(walkDirRec(root&"/"&line)).join("")):
-            removeDir(root&"/"&line)
+        for line in lines installedDir&"/"&actualPackage&"/list_files":
+            discard tryRemoveFile(root&"/"&line)
+
+            if isEmptyOrWhitespace(toSeq(walkDirRec(root&"/"&line)).join("")):
+                removeDir(root&"/"&line)
 
     if not ignoreReplaces:
-        let pkgreplaces = parse_runfile(installedDir&"/"&actualPackage).replaces
-        for i in pkgreplaces:
+        for i in pkg.replaces:
             if symlinkExists(installedDir&"/"&i):
                 removeFile(installedDir&"/"&i)
